requirements.txt
fastapi[standard]
langchain_openai
langchain_chroma
langchain 
langchain_core
langchain_community
pypdf
streamlit
python-dotenv
requests
uvicorn